Splitting metatest into files
