+++ lib/metatest.rb
@@ -1,0 +1,7 @@
+module Mumukit::Metatest
+end
+
+require_relative './metatest/errors'
+require_relative './metatest/checker'
+require_relative './metatest/framework'
+require_relative './metatest/identity_runner'
