+++ lib/metatest/errors.rb
@@ -1,0 +1,10 @@
+module Mumukit::Metatest
+  class Aborted < StandardError
+  end
+
+  class Errored < StandardError
+  end
+
+  class Failed < StandardError
+  end
+end
