+++ lib/metatest/identity_runner.rb
@@ -1,0 +1,7 @@
+module Mumukit::Metatest
+  class IdentityRunner
+    def run(it, _example)
+      it
+    end
+  end
+end
