+++ lib/qsim_server.rb
@@ -8,6 +8,8 @@ end
 
 require 'erb'
 
+require_relative './metatest'
 require_relative './test_hook'
 require_relative './metadata_hook'
+require_relative './checker'
 require_relative './html_renderer'
